ContentProvider functioning complete
ContentProvider functions as intended. The user is able to search
movies, the movies are stored in a file and the file is pulled to
display data in listview. The user can then filter the listview by year
in which the listview will change to only those items containing that
year. The user is also able to refresh their previous search or start a
new search.